requirements.txt
python-dotenv
scrapy
parsel
xmltodict
js2xml
lxml
